refactor: change login routes
