+++ src/app/app.component.ts
@@ -1,10 +1,10 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
-import { LoginComponent } from './auth/login/login.component';
 import { StatCardComponent } from './shared/components/stat-card/stat-card.component';
 import { SidebarComponent } from './shared/components/sidebar/sidebar.component';
 import { INavLinks } from './shared/components/sidebar/sidebar.interface';
 import { NavbarComponent } from './shared/components/navbar/navbar.component';
+import { LoginComponent } from './pages/auth/login/login.component';
 
 @Component({
   selector: 'app-root',
